widget関係のコードを別ファイルに保存＋見やすく整形
