+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:math';
 import 'dart:io' show Platform;
 
 import 'package:android_intent_plus/android_intent.dart';
 import 'package:chimimoryo_autumn/repository/repository.dart';
+import 'package:chimimoryo_autumn/widget/widget.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -12,40 +12,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:home_widget/home_widget.dart';
 import 'package:workmanager/workmanager.dart';
-
-/// Used for Background Updates using Workmanager Plugin
-void callbackDispatcher() {
-  Workmanager().executeTask((taskName, inputData) {
-    final now = DateTime.now();
-    return Future.wait<bool?>([
-      HomeWidget.saveWidgetData(
-        'title',
-        'Updated from Background',
-      ),
-      HomeWidget.saveWidgetData(
-        'message',
-        '${now.hour.toString().padLeft(2, '0')}:${now.minute.toString().padLeft(2, '0')}',
-      ),
-      HomeWidget.updateWidget(
-        name: 'HomeWidgetExampleProvider',
-        iOSName: 'HomeWidgetExample',
-      ),
-    ]).then((value) {
-      return !value.contains(false);
-    });
-  });
-}
-
-/// Called when Doing Background Work initiated from Widget
-void backgroundCallback(Uri? data) async {
-  if (data!.host == 'titleclicked') {
-    final greetings = 'こんにちは';
-
-    await HomeWidget.saveWidgetData<String>('title', greetings);
-    await HomeWidget.updateWidget(
-        name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample');
-  }
-}
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -86,91 +52,15 @@ class _MyHomePageState extends State<MyHomePage> {
   final TextEditingController _titleController = TextEditingController();
   final TextEditingController _messageController = TextEditingController();
 
-  // 将来ウィジェットにおける操作から何かを反映したい時用
   @override
   void initState() {
     super.initState();
     getLocationAndLaunchPay();
+
+    // 将来ウィジェットにおける操作から何かを反映したい時用
     HomeWidget.setAppGroupId('YOUR_GROUP_ID');
     HomeWidget.registerBackgroundCallback(backgroundCallback);
   }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    _checkForWidgetLaunch();
-    HomeWidget.widgetClicked.listen(_launchedFromWidget);
-  }
-
-  @override
-  void dispose() {
-    _titleController.dispose();
-    _messageController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _sendData() async {
-    try {
-      Future.wait([
-        HomeWidget.saveWidgetData<String>('title', _titleController.text),
-        HomeWidget.saveWidgetData<String>('message', _messageController.text),
-      ]);
-    } on PlatformException catch (exception) {
-      debugPrint('Error Sending Data. $exception');
-    }
-  }
-
-  Future<void> _updateWidget() async {
-    try {
-      HomeWidget.updateWidget(
-          name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample');
-    } on PlatformException catch (exception) {
-      debugPrint('Error Updating Widget. $exception');
-    }
-  }
-
-  // For load data in home widget
-  Future<void> _loadData() async {
-    try {
-      Future.wait([
-        HomeWidget.getWidgetData<String>('title', defaultValue: 'Default Title')
-            .then((value) => _titleController.text = value!),
-        HomeWidget.getWidgetData<String>('message',
-                defaultValue: 'Default Message')
-            .then((value) => _messageController.text = value!),
-      ]);
-    } on PlatformException catch (exception) {
-      debugPrint('Error Getting Data. $exception');
-    }
-  }
-
-  // For send information to Home Widget
-  Future<void> _sendAndUpdate() async {
-    await _sendData();
-    await _updateWidget();
-  }
-
-  void _checkForWidgetLaunch() {
-    HomeWidget.initiallyLaunchedFromHomeWidget().then(_launchedFromWidget);
-  }
-
-  /// For PBI「おまけ：ウィジェット起動時とアプリ起動時で異なる動きができるかを調査」 in PBI4
-  void _launchedFromWidget(Uri? uri) {
-    if (uri != null) {
-      showDialog(
-          context: context,
-          builder: (buildContext) => AlertDialog(
-                title: Text('App started from HomeScreenWidget'),
-                content: Text('Here is the URI: $uri'),
-              ));
-    }
-  }
-
-  final valueRate = {
-    "SevenEleven": {"Linepay": 2, "Paypay": 1},
-    "Lawson": {"Linepay": 2, "Paypay": 1},
-    "FamilyMart": {"Linepay": 2, "Paypay": 1},
-  };
 
   void getLocationAndLaunchPay() async {
     final store = getLocation();
@@ -249,4 +139,77 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
     );
   }
+
+  // ----------Widget関係（ここから）----------
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _checkForWidgetLaunch();
+    HomeWidget.widgetClicked.listen(_launchedFromWidget);
+  }
+
+  @override
+  void dispose() {
+    _titleController.dispose();
+    _messageController.dispose();
+    super.dispose();
+  }
+
+  Future<void> _sendData() async {
+    try {
+      Future.wait([
+        HomeWidget.saveWidgetData<String>('title', _titleController.text),
+        HomeWidget.saveWidgetData<String>('message', _messageController.text),
+      ]);
+    } on PlatformException catch (exception) {
+      debugPrint('Error Sending Data. $exception');
+    }
+  }
+
+  Future<void> _updateWidget() async {
+    try {
+      HomeWidget.updateWidget(
+          name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample');
+    } on PlatformException catch (exception) {
+      debugPrint('Error Updating Widget. $exception');
+    }
+  }
+
+  // For load data in home widget
+  Future<void> _loadData() async {
+    try {
+      Future.wait([
+        HomeWidget.getWidgetData<String>('title', defaultValue: 'Default Title')
+            .then((value) => _titleController.text = value!),
+        HomeWidget.getWidgetData<String>('message',
+                defaultValue: 'Default Message')
+            .then((value) => _messageController.text = value!),
+      ]);
+    } on PlatformException catch (exception) {
+      debugPrint('Error Getting Data. $exception');
+    }
+  }
+
+  // For send information to Home Widget
+  Future<void> _sendAndUpdate() async {
+    await _sendData();
+    await _updateWidget();
+  }
+
+  void _checkForWidgetLaunch() {
+    HomeWidget.initiallyLaunchedFromHomeWidget().then(_launchedFromWidget);
+  }
+
+  /// For PBI「おまけ：ウィジェット起動時とアプリ起動時で異なる動きができるかを調査」 in PBI4
+  void _launchedFromWidget(Uri? uri) {
+    if (uri != null) {
+      showDialog(
+          context: context,
+          builder: (buildContext) => AlertDialog(
+                title: Text('App started from HomeScreenWidget'),
+                content: Text('Here is the URI: $uri'),
+              ));
+    }
+  }
+  // ----------Widget関係（ここまで）----------
 }
