+++ lib/widget/widget.dart
@@ -1,0 +1,36 @@
+import 'package:home_widget/home_widget.dart';
+import 'package:workmanager/workmanager.dart';
+
+/// Used for Background Updates using Workmanager Plugin
+void callbackDispatcher() {
+  Workmanager().executeTask((taskName, inputData) {
+    final now = DateTime.now();
+    return Future.wait<bool?>([
+      HomeWidget.saveWidgetData(
+        'title',
+        'Updated from Background',
+      ),
+      HomeWidget.saveWidgetData(
+        'message',
+        '${now.hour.toString().padLeft(2, '0')}:${now.minute.toString().padLeft(2, '0')}',
+      ),
+      HomeWidget.updateWidget(
+        name: 'HomeWidgetExampleProvider',
+        iOSName: 'HomeWidgetExample',
+      ),
+    ]).then((value) {
+      return !value.contains(false);
+    });
+  });
+}
+
+/// Called when Doing Background Work initiated from Widget
+void backgroundCallback(Uri? data) async {
+  if (data!.host == 'titleclicked') {
+    const greetings = 'こんにちは';
+
+    await HomeWidget.saveWidgetData<String>('title', greetings);
+    await HomeWidget.updateWidget(
+        name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample');
+  }
+}
